requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.3
cloudinary==1.36.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.3
deprecation==2.1.0
Django==4.2.20
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-postgresql==0.0.3
django-rest-framework==0.1.0
django-storages==1.14
djangorestframework==3.16.0
drf-yasg==1.21.10
exceptiongroup==1.2.2
frozenlist==1.6.0
future==1.0.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==2.10.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gotrue==2.12.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
multidict==6.4.3
oauthlib==3.3.1
openapi-codec==1.3.2
packaging==25.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.32.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.3
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.5.0
websockets==14.2
whitenoise==6.9.0
yarl==1.20.0
django-cleanup==2.1.0